requirements.txt
a2wsgi==1.10.4
about-time==4.2.1
aiohttp==3.9.3
aioredis==2.0.1
alembic==1.13.1
apscheduler==3.10.4
argcomplete==3.2.3
asyncpg==0.29.0
black==24.4.1
cloudpickle==3.0.0
codecov==2.1.13
connexion==2.14.2
databases==0.9.0
dill==0.3.8
email-validator==2.1.1
execnet==2.0.2
fastapi==0.110.1
flask-caching==2.1.0
flask-session==0.7.0
gunicorn==21.2.0
httpx==0.27.0
PyJWT==2.8.0
lazy-object-proxy==1.10.0
linkify-it-py==2.0.3
loguru==0.7.2
markdown==3.6
marshmallow-oneofschema==3.1.1
mdit-py-plugins==0.4.0
mlxtend==0.23.1
openpyxl==3.1.2
passlib==1.7.4
pendulum==3.0.0
polars==0.20.17
prometheus-fastapi-instrumentator==7.0.0
psutil==5.9.8
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
pytest-asyncio==0.23.6
pytest-mock==3.14.0
pytest-watch==4.2.0
python-daemon==3.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-nvd3==0.15.0
rfc3339-validator==0.1.4
rich-argparse==1.4.0
sentry-sdk==2.1.1
setproctitle==1.3.3
sqlalchemy-jsonfield==1.0.2
sqlmodel==0.0.18
sqlparse==0.4.4
timy==0.4.2
toml==0.10.2
unicodecsv==0.14.1
uvicorn==0.15.0
